Fix integration test, parameterize buckets

+++ backend/src/test/java/io/zeitmaschine/image/ImagingIntegrationTest.java
@@ -22,11 +22,11 @@ import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 import io.minio.BucketExistsArgs;
 import io.minio.ListObjectsArgs;
-import io.minio.MakeBucketArgs;
 import io.minio.MinioClient;
 import io.minio.PutObjectArgs;
 import io.minio.RemoveBucketArgs;
@@ -42,9 +42,9 @@ import io.minio.errors.XmlParserException;
 import io.zeitmaschine.s3.S3Config;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ActiveProfiles("integration-test")
 public class ImagingIntegrationTest {
 
-    private static final String TEST_BUCKET = "test-bucket";
     private static final String TEST_IMAGE_NAME = "IMG_20181001_185137.jpg";
 
     @Autowired
@@ -54,6 +54,7 @@ public class ImagingIntegrationTest {
     private S3Config config;
 
     private MinioClient minioClient;
+    private String bucket;
 
     @BeforeEach
     void setUp() throws IOException, InvalidKeyException, InvalidResponseException, InsufficientDataException, NoSuchAlgorithmException, ServerException, InternalException, XmlParserException, InvalidBucketNameException, ErrorResponseException, RegionConflictException {
@@ -63,9 +64,16 @@ public class ImagingIntegrationTest {
 
         Resource image = new ClassPathResource("images/IMG_20181001_185137.jpg");
 
-        minioClient.makeBucket(MakeBucketArgs.builder().bucket(TEST_BUCKET).build());
+        this.bucket = config.getBucket();
+        String cacheBucket = config.getCacheBucket();
+
+        // Created in S3Repository
+        if (!minioClient.bucketExists(BucketExistsArgs.builder().bucket(bucket).build()) ||
+                !minioClient.bucketExists(BucketExistsArgs.builder().bucket(cacheBucket).build()))
+            throw new RuntimeException("Buckets not created.");
+
         minioClient.putObject(PutObjectArgs.builder()
-                .bucket(TEST_BUCKET)
+                .bucket(bucket)
                 .object(TEST_IMAGE_NAME)
                 .stream(image.getInputStream(), image.contentLength(), -1)
                 .contentType(MediaType.IMAGE_JPEG_VALUE)
@@ -75,24 +83,30 @@ public class ImagingIntegrationTest {
     @AfterEach
     void tearDown() throws IOException, InvalidKeyException, InvalidResponseException, InsufficientDataException, NoSuchAlgorithmException, ServerException, InternalException, XmlParserException, InvalidBucketNameException, ErrorResponseException {
         minioClient.listObjects(ListObjectsArgs.builder()
-                .bucket(TEST_BUCKET)
+                .bucket(config.getBucket())
                 .build())
                 .forEach(obj -> {
                     try {
                         minioClient.removeObject(RemoveObjectArgs.builder()
-                                .bucket(TEST_BUCKET)
+                                .bucket(bucket)
                                 .object(obj.get().objectName())
                                 .build());
                     } catch (Exception e) {
                         throw new RuntimeException("Failed to delete object.", e);
                     }
                 });
-        minioClient.removeBucket(RemoveBucketArgs.builder().bucket(TEST_BUCKET).build());
+        minioClient.removeBucket(RemoveBucketArgs.builder().bucket(bucket).build());
     }
 
+    /**
+     * This test ensures that an authenticated user can call the image endpoint with a request for an image
+     * operation. The passed image must exist in the s3 bucket, from where it is loaded and passed to the imaginary service.
+     *
+     * The expected result is a resized image.
+     */
     @Test
     @WithMockUser
-    void name() {
+    void routedImageOperation() {
         webTestClient.get()
                 .uri(uriBuilder -> uriBuilder
                         .path("image/")
